Revert "feat : PlayPause(), backTenSecondBefore(), goTenSecondLater(), handleNext( go to next music ), handlePrevious( back to music before ) and navigation seekbar functions"
This reverts commit 56629ac3a31dbdaea625e4223801b9c786bb12f0.
