+++ frontend/dwaves-player/package.json
@@ -1,27 +1,27 @@
 {
-  "name": "dwaves-app",
+  "name": "dwaves-player",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@mantine/core": "^5.5.5",
-    "@mantine/hooks": "^5.5.1",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
+    "@heroicons/react": "^1.0.6",
+    "@mantine/core": "^4.2.12",
+    "@mantine/hooks": "^4.2.12",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
-    "@types/howler": "^2.2.7",
     "@types/jest": "^27.5.2",
-    "@types/node": "^16.11.64",
-    "@types/react": "^18.0.21",
+    "@types/node": "^16.11.43",
+    "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
-    "@types/react-router-dom": "^5.3.3",
     "howler": "^2.2.3",
+    "node-sass": "^7.0.1",
     "react": "^18.2.0",
-    "react-audio-player": "^0.17.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.4.1",
+    "react-howler": "^5.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
-    "sass": "^1.55.0",
-    "typescript": "^4.8.4",
+    "spinners-react": "^1.0.7",
+    "typescript": "^4.7.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -47,5 +47,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/howler": "^2.2.7"
   }
 }
