+++ frontend/dwaves-player/src/App.scss
@@ -2,6 +2,5 @@ $dark : #3C4A68;
 
 .App {
     background-color: $dark;
-    overflow-x: hidden;
 }
