+++ frontend/dwaves-player/src/App.tsx
@@ -8,6 +8,7 @@ import {
 
 import { Player } from './Pages/Player';
 import { Loader } from './Components/Loader';
+import { Divider } from '@mantine/core';
 
 
 function App() {
