--- frontend/dwaves-player/src/Musics/datasongs.json
@@ -1,44 +1,0 @@
-[
-    {
-        "Artist": "Dinos",
-        "Title": "diptyque",
-        "featuring": "",
-        "likes": 190, 
-        "progress": 0,
-        "length": 0,
-        "UsersLiked": [
-                { "Name": "", "Avatar": "http://via.placeholder.com/360x360" }
-            ],
-        "Src": "http://localhost:3000/audiotest1.mp3",
-        "Cover": "http://localhost:3000/stamina1.jpg",
-        "outDate": ""
-    },
-    {
-        "Artist": "Dinos",
-        "Title": "Moins un",
-        "featuring": "Nekfeu",
-        "likes": 190, 
-        "progress": 0,
-        "length": 0,
-        "UsersLiked": [
-                { "Name": "", "Avatar": "http://via.placeholder.com/360x360" }
-            ],
-        "Src": "http://localhost:3000/audiotest2.mp3",
-        "Cover": "http://localhost:3000/stamina2.jpg",
-        "outDate": ""
-    },
-    {
-        "Artist": "Dinos",
-        "Title": "Paranoiaque",
-        "featuring": "",
-        "likes": 190, 
-        "progress": 0,
-        "length": 0,
-        "UsersLiked": [
-                { "Name": "", "Avatar": "http://via.placeholder.com/360x360" }
-            ],
-        "Src": "http://localhost:3000/audiotest3.mp3",
-        "Cover": "http://localhost:3000/stamina3.jpg",
-        "outDate": ""
-    }
- ]
