+++ frontend/dwaves-player/src/Styles/Player.scss
@@ -2,7 +2,7 @@
     position: relative;
     height: 100vh;
     width: 100vw;
-    background-image: url("http://localhost:3000/stamina1.jpg");
+    background-image: url("../Images/stamina.jpg");
     background-attachment: fixed;
     background-size: cover;
     background-position: center;
@@ -130,18 +130,4 @@
         border-radius: 100px;
     }
 
-}
-
-.seekbar{
-    margin-top: 10%;
-    width: 100%;
-    height: 5px;
-    background: #000;
-    border-radius: 10px;
-    .time{
-        height: 100%;
-        width: 0;
-        background: blue;
-        border-radius: 10px;
-    }
 }
